requirements.txt
aiohttp
camelot-py[cv]
Pillow
pdf2image
starlette
fastapi
python-multipart
uvicorn
aiofiles
pandas
colorama
